Outsource GeopointParser. (see #58)
