+++ src/cgeo/geocaching/geopoint/Geopoint.java
@@ -1,8 +1,5 @@
 package cgeo.geocaching.geopoint;
 
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-
 /**
  * Abstraction of geographic point.
  */
@@ -12,12 +9,6 @@ public class Geopoint
     public static final double deg2rad   = Math.PI / 180;
     public static final double rad2deg   = 180 / Math.PI;
     public static final float  erad      = 6371.0f;
-
-    private enum LatLon
-    {
-        LAT,
-        LON
-    }
     
     private double latitude;
     private double longitude;
@@ -47,12 +38,12 @@ public class Geopoint
      * Creates new Geopoint with latitude and longitude parsed from string.
      *
      * @param text string to parse
-     * @see parse()
+     * @see GeopointParser.parse()
      */
     public Geopoint(final String text)
     {
-        setLatitude(parseLatitude(text));
-        setLongitude(parseLongitude(text));
+        setLatitude(GeopointParser.parseLatitude(text));
+        setLongitude(GeopointParser.parseLongitude(text));
     }
 
     /**
@@ -63,129 +54,6 @@ public class Geopoint
     public Geopoint(final Geopoint gp)
     {
         this(gp.getLatitude(), gp.getLongitude());
-    }
-
-    /**
-     * Parses a pair of coordinates (latitude and longitude) out of a String.
-     * Accepts following formats and combinations of it:
-     *      X DD
-     *      X DD°
-     *      X DD° MM
-     *      X DD° MM.MMM
-     *      X DD° MM SS
-     *
-     * as well as:
-     *      DD.DDDDDDD
-     *
-     * Both . and , are accepted, also variable count of spaces (also 0)
-     *
-     * @param text the string to parse
-     * @return an Geopoint with parsed latitude and longitude
-     * @throws ParseException if lat or lon could not be parsed
-     */
-    public static Geopoint parse(final String text)
-    {
-        
-        double lat = parseLatitude(text);
-        double lon = parseLongitude(text);
-        
-        return new Geopoint(lat, lon);
-    }
-    
-    /* (non JavaDoc)
-     * Helper for coordinates-parsing.
-     */
-    private static double parseHelper(final String text, final LatLon latlon)
-    {
-        final Pattern patternLat = Pattern.compile("([NS])\\s*(\\d+)°?(\\s*(\\d+)([\\.,](\\d+)|'?\\s*(\\d+)(''|\")?)?)?");
-        final Pattern patternLon = Pattern.compile("([WE])\\s*(\\d+)°?(\\s*(\\d+)([\\.,](\\d+)|'?\\s*(\\d+)(''|\")?)?)?");
-        final Pattern patternDec = Pattern.compile("^(-?\\d+([\\.,]\\d+)?)\\s*(-?\\d+([\\.,]\\d+)?)?$");
-
-        Matcher matcher;
-
-        if (LatLon.LAT == latlon)
-        {
-            matcher = patternLat.matcher(text);
-        }
-        else
-        {
-            matcher = patternLon.matcher(text);
-        }
-
-        if (matcher.find())
-        {
-            int sign      = 1;
-            int degree    = 0;
-            int minutes   = 0;
-            int seconds   = 0;
-            
-            if (matcher.group(1).equalsIgnoreCase("S") || matcher.group(1).equalsIgnoreCase("W"))
-            {
-                sign = -1;
-            }
-            
-            degree = Integer.parseInt(matcher.group(2));
-            
-            if (null != matcher.group(4))
-            {
-                minutes = Integer.parseInt(matcher.group(4));
-                
-                if (null != matcher.group(6))
-                {
-                    seconds = Math.round(Float.parseFloat("0." + matcher.group(6)) * 60);
-                }
-                else if (null != matcher.group(7))
-                {
-                    seconds = Integer.parseInt(matcher.group(7));
-                }
-            }
-            
-            return (double) sign * ((double) degree + (double) minutes / 60 + (double) seconds / 3600);
-        }
-        else // Nothing found with "N 52...", try to match string as decimaldegree
-        {
-            matcher = patternDec.matcher(text);
-
-            if (matcher.find())
-            {
-                if (LatLon.LAT == latlon)
-                {
-                    return Double.parseDouble(matcher.group(1).replaceAll(",", "."));
-                }
-                else if (null != matcher.group(3))
-                {
-                    return Double.parseDouble(matcher.group(3).replaceAll(",", "."));
-                }
-            }
-        }
-        
-        throw new ParseException("Could not parse coordinates as " + latlon + ": \"" + text + "\"");
-    }
-
-    /**
-     * Parses latitude out of a given string.
-     *
-     * @see parse()
-     * @param text the string to be parsed
-     * @return the latitude as decimaldegree
-     * @throws ParseException if latitude could not be parsed
-     */
-    public static double parseLatitude(final String text)
-    {
-        return parseHelper(text, LatLon.LAT);
-    }
-
-    /**
-     * Parses longitude out of a given string.
-     *
-     * @see parse()
-     * @param text the string to be parsed
-     * @return the longitude as decimaldegree
-     * @throws ParseException if longitude could not be parsed
-     */
-    public static double parseLongitude(final String text)
-    {
-        return parseHelper(text, LatLon.LON);
     }
 
     /**
@@ -233,7 +101,7 @@ public class Geopoint
      */
     public Geopoint setLatitude(final String lat)
     {
-        return setLatitude(parseLatitude(lat));
+        return setLatitude(GeopointParser.parseLatitude(lat));
     }
 
     /**
@@ -301,7 +169,7 @@ public class Geopoint
      */
     public Geopoint setLongitude(final String lon)
     {
-        return setLongitude(parseLongitude(lon));
+        return setLongitude(GeopointParser.parseLongitude(lon));
     }
 
     /**
@@ -518,17 +386,6 @@ public class Geopoint
         private static final long serialVersionUID = 1L;
 
         public GeopointException(String msg)
-        {
-            super(msg);
-        }
-    }
-
-    public static class ParseException
-        extends GeopointException
-    {
-        private static final long serialVersionUID = 1L;
-
-        public ParseException(String msg)
         {
             super(msg);
         }
